Add wrapping functionality to the new north/south/east/west movement
This adds wrapping functionality by "moving" windows that are on the 
right axis but in the wrong direction to the correct side of the movement. 
The windows are moved by either adding or subtracting the window height/width.

Maybe the IsWindowInDirection function should be renamed. I don't really know
if IsWindowOnAxis is a better name.